Making refactor to Queries and Mutations Sctructure
